Modify project structure, add cross-platform support.
