+++ cmake/ccache.cmake
@@ -1,7 +1,1 @@
-find_program(CCACHE_PROGRAM ccache)
-if(CCACHE_PROGRAM)
-    set_property(GLOBAL PROPERTY RULE_LAUNCH_COMPILE "${CCACHE_PROGRAM}")
-    message(STATUS "ccache: found")
-else()
-    message(STATUS "ccache: not found")
-endif()
+
